+++ NEWS.md
@@ -10,6 +10,7 @@ MAJOR CHANGES
 
 MINOR FEATURES AND BUG FIXES 
 
+* fixed warning when subset.rspec was provided with a logical vector
 * fixed harmless warning when summary.colspace() was used on a tcs object
 * `by` argument in merge.rspec() is no longer ignored
 * fixed bug in voloverlap() when plot = TRUE